+++ spring-data/src/main/java/com/epam/core/dto/EventDto.java
@@ -3,6 +3,7 @@ package com.epam.core.dto;
 import com.epam.core.model.Event;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -10,5 +11,5 @@ public class EventDto implements Event {
     private long id;
     private String title;
     private Date date;
-    private int ticketPrice;
+    private BigDecimal ticketPrice;
 }
