+++ spring-data/src/main/java/com/epam/core/dto/UserAccountDto.java
@@ -3,9 +3,11 @@ package com.epam.core.dto;
 import com.epam.core.model.UserAccount;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UserAccountDto implements UserAccount {
     private long id;
     private long userId;
-    private int balance;
+    private BigDecimal balance;
 }
