+++ spring-data/src/main/java/com/epam/core/model/Event.java
@@ -1,5 +1,6 @@
 package com.epam.core.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -16,6 +17,6 @@ public interface Event {
     void setTitle(String title);
     Date getDate();
     void setDate(Date date);
-    int getTicketPrice();
-    void setTicketPrice(int ticketPrice);
+    BigDecimal getTicketPrice();
+    void setTicketPrice(BigDecimal ticketPrice);
 }
