+++ spring-data/src/main/java/com/epam/core/model/UserAccount.java
@@ -1,10 +1,12 @@
 package com.epam.core.model;
 
+import java.math.BigDecimal;
+
 public interface UserAccount {
     long getId();
     void setId(long id);
     long getUserId();
     void setUserId(long userId);
-    int getBalance();
-    void setBalance(int balance);
+    BigDecimal getBalance();
+    void setBalance(BigDecimal balance);
 }
