+++ spring-data/src/main/resources/db/create_events_table.sql
@@ -4,5 +4,5 @@ create table events (
     id serial primary key,
     title varchar(30),
     date date,
-	ticket_price integer
+	ticket_price numeric(20, 2)
 );
